added an advanced header file

+++ main.cpp
@@ -1,8 +1,8 @@
-#include<iostream>
+#include<bits/stdc++.h>
 using namespace std;
 
 int main(){
-    int n;
+    long long int n;
     cin>>n;
     for(int i=1; i<=n; i++)
     cout<<i<<endl;
